+++ addons/survey/static/tests/tours/survey_chained_conditional_questions.js
@@ -2,7 +2,6 @@
 
 import { queryAll } from "@odoo/hoot-dom";
 import { registry } from "@web/core/registry";
-import { TourError } from "@web_tour/tour_service/tour_utils";
 
 registry.category("web_tour.tours").add('test_survey_chained_conditional_questions', {
     test: true,
@@ -97,6 +96,6 @@ registry.category("web_tour.tours").add('test_survey_chained_conditional_questio
 
 export function expectHiddenQuestion (questionTitle, msg){
     if (queryAll(`div.js_question-wrapper.d-none:contains('${questionTitle}')`).length !== 1) {
-        throw new TourError(msg);
+        console.error(msg);
     }
 }
