+++ addons/website/static/tests/tours/media_dialog.js
@@ -1,6 +1,5 @@
 /** @odoo-module */
 
-import { TourError } from "@web_tour/tour_service/tour_utils";
 import wTourUtils from '@website/js/tours/tour_utils';
 
 wTourUtils.registerWebsitePreviewTour("website_media_dialog_undraw", {
@@ -69,7 +68,7 @@ wTourUtils.registerWebsitePreviewTour("website_media_dialog_external_library", {
             const imgName = this.$anchor[0].querySelector(selector).title;
             const uploadedImgs = this.$anchor[0].querySelectorAll(`${selector}[title='${imgName}']`);
             if (uploadedImgs.length !== 1) {
-                throw new TourError(`${uploadedImgs.length} attachment(s) were found. Exactly 1 should have been created.`);
+                console.error(`${uploadedImgs.length} attachment(s) were found. Exactly 1 should have been created.`);
             }
         },
     },
