+++ addons/website_sale_autocomplete/static/tests/autocomplete_tour.js
@@ -1,12 +1,11 @@
 /** @odoo-module */
 
 import { registry } from "@web/core/registry";
-import { TourError } from "@web_tour/tour_service/tour_utils";
 import tourUtils from '@website_sale/js/tours/tour_utils';
 
 
 function fail (errorMessage) {
-    throw new TourError(errorMessage);
+    console.error(errorMessage);
 }
 
 registry.category("web_tour.tours").add('autocomplete_tour', {
